requirements.txt
apps==0.0.0.0.dev20200918
asgiref==3.3.4
attrs==21.2.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-perms==2.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
